requirements.txt
flask
werkzeug
requests
feedgen @ git+https://github.com/yimingliu/python-feedgen@main
python-dateutil
parsel
sentry-sdk[flask]
cloudscraper

